+++ laravel/database/eloquent/relationships/relationship.php
@@ -51,7 +51,7 @@ abstract class Relationship extends Query {
 		// Next we'll set the fluent query builder for the relationship and
 		// constrain the query such that it only returns the models that
 		// are appropriate for the relationship.
-		$this->table = $this->query();
+		$this->table = $this->table();
 
 		$this->constrain();
 	}
@@ -96,7 +96,7 @@ abstract class Relationship extends Query {
 	 *
 	 * @return string
 	 */
-	protected function foreign_key()
+	public function foreign_key()
 	{
 		return static::foreign($this->base, $this->foreign);
 	}